% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkgrid.R
\name{mkDKN}
\alias{mkDKN}
\title{mk DKN grid}
\usage{
mkDKN(ext, res)
}
\arguments{
\item{ext}{A spatial layer that you want to have grided.}

\item{res}{Resolution of grid in meters.}
}
\description{
This function create a DKN-grid with the parameters: extent, resolution (in meters)
}
